caffeine: Consolidate set_error macro
