+++ plugins/caffeine/caffeine-log.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <util/dstr.h>
+
 #ifndef CAFFEINE_LOG_TITLE
 #error "Define CAFFEINE_LOG_TITLE before including caffeine-log.h"
 #endif
@@ -13,3 +15,13 @@
 #define log_debug(format, ...)  do_log(LOG_DEBUG, format, ##__VA_ARGS__)
 
 #define trace() log_debug("%s", __func__)
+
+#define set_error(output, fmt, ...) \
+	do { \
+		struct dstr message; \
+		dstr_init(&message); \
+		dstr_printf(&message, (fmt), ##__VA_ARGS__); \
+		log_error("%s", message.array); \
+		obs_output_set_last_error((output), message.array); \
+		dstr_free(&message); \
+	} while(false)
